### 2018-07-01 #20 @whedon
[PRE REVIEW]: CFD Python: the 12 steps to Navier-Stokes equations
**Submitting author:** @labarba (<a href="http://orcid.org/0000-0001-5812-2711">Lorena A. Barba</a>)
**Repository:** <a href="https://github.com/barbagroup/CFDPython" target ="_blank">https://github.com/barbagroup/CFDPython</a>
**Version:** v1.0
