### 2018-07-01 8ee1115
add worked-example papers
