+++ paper.bib
@@ -2,12 +2,35 @@
 %% http://bibdesk.sourceforge.net/
 
 
-%% Created for Lorena A Barba at 2018-06-30 21:49:32 -0400 
+%% Created for Lorena A Barba at 2018-07-01 11:19:20 -0400 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@article{chen2015worked,
+	Author = {Chen, Ouhao and Kalyuga, Slava and Sweller, John},
+	Date-Added = {2018-07-01 15:18:29 +0000},
+	Date-Modified = {2018-07-01 15:19:04 +0000},
+	Doi = {10.1037/edu0000018},
+	Journal = {Journal of Educational Psychology},
+	Number = {3},
+	Pages = {689},
+	Publisher = {American Psychological Association},
+	Title = {The worked example effect, the generation effect, and element interactivity.},
+	Volume = {107},
+	Year = {2015}}
+
+@article{sweller2006worked,
+	Author = {Sweller, John},
+	Date-Added = {2018-07-01 15:15:07 +0000},
+	Date-Modified = {2018-07-01 15:15:42 +0000},
+	Doi = {10.1016/j.learninstruc.2006.02.005},
+	Journal = {Learning and instruction},
+	Publisher = {Elsevier Science},
+	Title = {The worked example effect and human cognition.},
+	Year = {2006}}
 
 @book{rossant2018ipython,
 	Author = {Rossant, Cyrille},
