### 2018-07-01 #20 @labarba
Maybe one of these gentle humans would like to volunteer a review?
@petebachant @nicoguaro @leouieda @wigging @franktoffel @AlexS12
(picked among CFD Python stargazers on GitHub)
