### 2018-07-01 #20 @petebachant
I'd be willing to review as well.
