### 2018-07-02 #21 @whedon
[REVIEW]: CFD Python: the 12 steps to Navier-Stokes equations
**Submitting author:** @labarba (<a href="http://orcid.org/0000-0001-5812-2711">Lorena A. Barba</a>)
**Repository:** <a href="https://github.com/barbagroup/CFDPython" target ="_blank">https://github.com/barbagroup/CFDPython</a>
**Version:** v1.0

- [x] **Version:** Does the release version given match the repository release (v1.0)?

- [x] Does it describe how it has been used in the classroom or other settings, and how someone might adotp it?

- [x] **Version:** Does the release version given match the repository release (v1.0)?

- [x] Does it describe how it has been used in the classroom or other settings, and how someone might adotp it?
