### 2018-07-02 barbagroup/CFDPython#43 @petebachant
Missing release v1.0
For https://github.com/openjournals/jose-reviews/issues/21.
