### 2018-07-03 barbagroup/CFDPython#44 @nicoguaro
Typos in paper.md
For openjournals/jose-reviews#21.
- Third paragraph in section "Statement of Need": usting >> using
