### 2018-07-03 #21 @nicoguaro
@kyleniemeyer, do we expect Statement of need to be explicitly written in the README?
