### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
Comments on notebooks
For openjournals/jose-reviews#21.
