### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
## 01_Step_1

``ipython notebook`` >> ``jupyter notebook``
