### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
## 02_Step_2

- Cell [1],
  ``import numpy                 #we're importing numpy and calling it np locally``
  NumPy is called ``numpy`` and not ``np``
