### 2018-07-03 9852ce7
Fix comments and some renaming for JOSE submission
For openjournals/jose-reviews#21
