### 2018-07-03 barbagroup/CFDPython#43 @gforsyth
I think it makes sense to hold off on cutting a release until we resolve the other issues from review -- that way the tag position can reflect the paper most accurately.
