### 2018-07-03 247af78
Fix typo in `paper.md`
Resolves #44

For openjournals/jose-reviews#21
