### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
## 05_Step_4

- Cell [3], I think that it would be more consistent to use ``sympy.pi`` instead of ``numpy.py``
  ``phi = (sympy.exp(-(x - 4 * t)**2 / (4 * nu * (t + 1))) +
       sympy.exp(-(x - 4 * t - 2 * numpy.pi)**2 / (4 * nu * (t + 1))))`` >> ``phi = (sympy.exp(-(x - 4 * t)**2 / (4 * nu * (t + 1))) +
       sympy.exp(-(x - 4 * t - 2 * sympy.pi)**2 / (4 * nu * (t + 1))))``
