### 2018-07-03 #21 @kyleniemeyer
@nicoguaro It doesn't have to be called out explicitly by that name, I think, but the need should be written clearly in the README or docs
