### 2018-07-03 #21 @labarba
hmmm... I'm not sure it needs to go in the README? It needs to go in the *paper*.
