### 2018-07-03 #21 @nicoguaro
@labarba, according to my interpretation of the guidelines it should be in the README. And yes, also in the paper … if we follow the checklist.
