### 2018-07-03 #21 @labarba
That's pretty funny, because I wrote the guidelines 😬
