### 2018-07-03 barbagroup/CFDPython#43 @petebachant
Makes sense. Just wanted to throw up an issue since it's part of the review.
