### 2018-07-03 #21 @labarba
Do you all think it's useful to have such content also in the README? I see the paper as the "advertisement" of the work (à la Claerbout), and the README as instructions to the user.
