### 2018-07-03 #21 @labarba
Bear in mind, the journal is new and unique, and we're defining our genre in these very conversations.
