### 2018-07-03 #21 @nicoguaro
I think that it is OK if it just appears in the paper. But it appears twice in the checklist as well, and I am following point by point in the review. Maybe it is better to update the checklist.
