### 2018-07-03 #21 @labarba
You're right! The checklists were adapted from those for JOSS.

Our only accepted JOSE paper does not include that content in the README:
https://github.com/gboeing/pynamical

@kyleniemeyer Help! What do we do?
