### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
## 06_Array_Operations_with_NumPy

- After Cell [6]: it might be useful to mention that the running time depends on the computer. For some people it might be obvious, but that's not always the case.
