### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
## 13_Step_10

- Before cell [1], I would change ``b_{i,j}=-100$ at $i=nx*3/4, j=3/4 ny`` for ``b_{i,j}=-100$ at $i=3/4 nx, j=3/4 ny``
