### 2018-07-03 #21 @kyleniemeyer
@labarba why don't we move this conversation elsewhere for now, so we can resolve where we want this content? @nicoguaro it sounds like the paper at least satisfies this item, so let's hold off on the README question for now
