### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
## 14_Optimizing_Loops_with_Numba

- The notebook mixes spaces and tabs.
- ``laplace2d`` functions mix the alias for NumPy, using ``numpy`` and ``np``. Due to this  the notebook does not work, the following is the returned error

  > NameError: name 'np' is not defined
