### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
## 15_Step_11

It might be a good idea to consider streamplots besides glyphs for the vector visualization. It suffices to change

    pyplot.quiver(X, Y, u, v)

for

    pyplot.streamplot(X, Y, u, v)

For some vector fields this is easier to understand, and introduces another visualization technique.

This comment is a matter of taste, so I understand if the authors don't consider it needed.
