### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
## 18_Burgers_equation

- Cell [6] returns the following error

  > from JSAnimation.IPython_display import display_animation

  JSAnimation is not mentioned as a dependency. A workaround that is compatible with ``FuncAnimation```is to use the IPython magic ``%matplotlib notebook`` instead of ``%matplotlib inline`
