### 2018-07-03 barbagroup/CFDPython#45 @nicoguaro
## 19_Odd_Even_Decoupling

The notebook depends on a folder with videos that is missing from the repository.
