### 2018-07-03 #21 @nicoguaro
I reviewed the paper/module and I think that it suits the guidelines of JOSE. I have made some comments on the notebooks on barbagroup/CFDPython#45. There are three boxes that I have not marked in the checklist. Following my comments.

### Version

There are different notebooks that have different version numbers. Also, there is not a specific number version given in the repository. This has been mentioned by @petebachant in barbagroup/CFDPython#43.

### Installation instructions

There are specific instruction on how to install most dependencies, except for ``numbapro`` that seems to be deprecated. Also, I think that something extra needs to be done to get ``JSAnimation`` to works as a submodule, maybe something along the lines of (?):

    git submodule update --init --recursive

It would not hurt to specify that besides installing dependencies the module needs to be cloned.

### Content quality

Some of the "bonus notebooks" need some improvements. See barbagroup/CFDPython#45 for more details.
