### 2018-07-04 barbagroup/CFDPython#45 @gforsyth
Notebooks 17, 18 and 19 were never really edited or tested -- at this point my inclination is to just delete them (or move them to a subdirectory).  Especially since numbapro no longer exists and the numba stuff in general is all very out of date.

Thoughts, @labarba ?
