### 2018-07-04 barbagroup/CFDPython#45 @gforsyth
@labarba -- I've pushed up individual commits addressing the comments on notebooks 1, 2, 5, 6 and 13, respectively, in #46 .  If you disagree with any of the fixes just let me know and I can pull out the individual commits in question.

I like the suggestion about using `streamplot` (which, if I recall, was less reliable back when we wrote these).

Here's the existing quiver plot:
![quiver](https://user-images.githubusercontent.com/3596999/42281425-7df9d92e-7f71-11e8-8cae-6f41d74f34c7.png)
And here's the same plot using `streamplot` instead.
![streamplot](https://user-images.githubusercontent.com/3596999/42281437-860ee9ec-7f71-11e8-918c-903c6e661546.png)

Preference?
