### 2018-07-04 barbagroup/CFDPython#45 @labarba
Regarding notebooks 17, 18, 19 — these are drafts, really… we could move them to a WiP folder?
