### 2018-07-10 barbagroup/CFDPython#50 @petebachant
Novices may need more instructions on how to use notebooks
For https://github.com/openjournals/jose-reviews/issues/21.

The current instructions simply tell users to run the notebook server, but they should probably also include cloning the repository, then moving into the directory, running the server, then opening the first notebook.
