### 2018-07-10 #21 @petebachant
Regarding the review checklist, there's a typo in "Does it describe how it has been used in the classroom or other settings, and how someone might **adotp** it?
