### 2018-07-10 barbagroup/CFDPython#52 @petebachant
Statement of need could be more clear about the target audience
The statement of need in `paper.md` does a good job describing the effectiveness of the lessons, but doesn't clearly explain who would benefit from these lessons and why. For example, is this geared towards graduate students in mechanical engineering, physics, math, etc.?

>A statement of need: Do the authors clearly state the need for this module and who the target audience is?

https://github.com/openjournals/jose-reviews/issues/21
