### 2018-07-10 barbagroup/CFDPython#50 @petebachant
The README does specify that users should be using fresh notebooks, so they can simply view the nbviewer links in there instead of cloning the repo, but maybe instructions showing how to create a new directory, then opening a fresh notebook for each part of the lesson would be helpful.
