### 2018-07-10 barbagroup/CFDPython#43 @petebachant
Notebooks 0 and 1 have version numbers in them as well. These should probably be automatically inserted using `git describe`, if they're even necessary at all.
