### 2018-08-16 #21 @kyleniemeyer
Hi @petebachant @nicoguaro, sorry for disappearing for a while... are you satisfied with the changes made by the authors?
