### 2018-08-20 #21 @petebachant
The paper should be good to go after the final release is tagged (https://github.com/barbagroup/CFDPython/issues/43).
