### 2018-08-21 #21 @kyleniemeyer
thanks @nicoguaro @petebachant. I'm going to take a final look at the paper itself before accepting.
