### 2018-08-25 #21 @labarba
We do the release with all the revisions up to acceptance, at that point we do a Zenodo deposit and get a DOI, we give the DOI here in a comment for the editor to run the command `@whedon set <doi> as archive`. At that point, publication can proceed. (So the link to the JOSE published paper will need to be added in the README after publication.)
