### 2018-08-26 #21 @labarba
One more thing: the version gets captured in the Zenodo metadata if we follow this
https://guides.github.com/activities/citable-code/
