### 2018-08-26 #21 @kyleniemeyer
@gforsyth @labarba sorry for the delay—I have some suggestions and corrections for the paper:

- In the first paragraph, the sentence "This practical module takes students through 12 steps, one by one, at the end of which they will have programmed a solver for the two-dimensional Navier-Stokes equations, using finite differences." sounds a little off. Perhaps something like "This practical module takes students through 12 steps one by one; by the end, they will have programmed a solver for the two-dimensional Navier-Stokes equations using finite differences."
- 2nd para: "Guiding the students through these steps (without skipping any), they learn many valuable lessons." -> something like "Students learn many valuable lessons as the modules guides them through these steps (without skipping any)."
- same paragraph: "numerical diffusion, accuracy and convergence" is missing an Oxford comma (not sure if we have decided whether JOSE follows this [correct] convention, but I do 😄... plus I think the rest of the paper uses this)
- In the statement of need 1st paragraph, the sentence "This learning module places emphasis on practical experience with _programming_ the solution to fundamental mathematical models that can represent fluid behavior." seems to be missing something; maybe "... programming *as* the solution..." ?
- Throughout the paper: please replace the hyphen in "Navier–Stokes" with an en-dash (`--`). Same for Courant-Friedrichs-Lewy on the second page.
- Third paragraph in statement of need: "As they become perplexed of this behavior..." -> "As they become perplexed at this behavior...". Also, missing comma in "consistency, stability and convergence".
- In the fourth paragraph of the statement of need, there are a number of inline URLs. Should any (or all) of these instead be proper citations?
- In the last paragraph, the second citation "Chen, Kalyuga, and Sweller (2015)" needs to use a different cite command to avoid the doubled parentheses.
- Reference Ketcheson 2014 is missing some information; here is the full BibTeX entry:
```bibtex
@InProceedings{ ketcheson-proc-scipy-2014,
  author    = { {D}avid {I}. {K}etcheson },
  title     = { {T}eaching numerical methods with {I}{P}ython notebooks and inquiry-based learning },
  booktitle = { {P}roceedings of the 13th {P}ython in {S}cience {C}onference },
  pages     = { 19 - 25 },
  year      = { 2014 },
  editor    = { {S}t\'efan van der {W}alt and {J}ames {B}ergstra }
}
```
- Reference Sweller 2006 is missing some information (volume, issue, pages)
