### 2018-10-30 a99f553
fix broken links to nbviewer
