### 2018-11-12 33b1a9d
edit JOSE paper
