+++ paper.bib
@@ -28,8 +28,12 @@
 	Date-Modified = {2018-07-01 15:15:42 +0000},
 	Doi = {10.1016/j.learninstruc.2006.02.005},
 	Journal = {Learning and instruction},
+	Volume = {16},
+	Number = {2},
+	Pages = {165--169},
+	Doi = {10.1016/j.learninstruc.2006.02.005},
 	Publisher = {Elsevier Science},
-	Title = {The worked example effect and human cognition.},
+	Title = {The worked example effect and human cognition},
 	Year = {2006}}
 
 @book{rossant2018ipython,
@@ -42,9 +46,12 @@
 
 @inproceedings{ketcheson2014-scipy,
 	Author = {Ketcheson, David I.},
-	Booktitle = {Proceedings of the 13th SciPy Conferenece},
 	Date-Added = {2018-07-01 01:44:36 +0000},
 	Date-Modified = {2018-07-01 01:47:12 +0000},
 	Note = {http://conference.scipy.org/proceedings/scipy2014/pdfs/ketcheson.pdf},
 	Title = {Teaching numerical methods with {IPython} notebooks and inquiry-based learning},
-	Year = {2014}}
+	Year = {2014},
+    booktitle = { {P}roceedings of the 13th {P}ython in {S}cience {C}onference },
+    pages     = { 19 -- 25 },
+    url = {http://hdl.handle.net/10754/346689},
+    editor    = { {S}t\'efan van der {W}alt and {J}ames {B}ergstra }}
