### 2018-11-12 #21 @labarba
@nicoguaro, @petebachant — Version will be updated with a release and archive in Zenodo.

@nicoguaro — Apart from the version, you do have a couple of unticked boxes, even though you said above that you're ready to recommend publication.

@kyleniemeyer — I've updated the paper with your editorial suggestions.

@gforsyth — Could you make a release and Zenodo deposit? Thanks!
