+++ requirements.txt
@@ -1,0 +1,6 @@
+ipywidgets==7.2.1
+jupyter==1.0.0
+numpy==1.14.3
+matplotlib==2.2.2
+scipy==1.1.0
+sympy==1.1.1
