### 2018-11-12 #21 @kyleniemeyer
@arfon can you take a look at the DOI in the Crossref deposit XML? It looks like the DOI isn't actually present: https://github.com/openjournals/jose-papers/pull/5/files
