### 2018-11-12 #21 @kyleniemeyer
@arfon nvm, I compared against JOSS articles and it looks the same.
