### 2018-11-12 #21 @kyleniemeyer
@labarba @gforsyth congrats, your paper is published!

@nicoguaro, @petebachant: thanks so much with your help reviewing this submission!
